% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{Compilation}
\alias{Compilation}
\title{Snaptron compilation enum}
\format{An object of class \code{environment} of length 17.}
\usage{
Compilation
}
\description{
Snaptron compilation enum
}
\keyword{datasets}
