% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{Coordinates}
\alias{Coordinates}
\title{Enum for Snaptron Coordinate modifiers}
\format{An object of class \code{environment} of length 4.}
\usage{
Coordinates
}
\description{
Enum for Snaptron Coordinate modifiers
}
\section{Fields}{

\describe{
\item{\code{Exact}}{Return junctions whose start and end
coordinates are match the boundaries of the
region requested.}

\item{\code{Within}}{Return junctions whose start and end
coordinates are within the boundaries of the
region requested.}

\item{\code{StartIsExactOrWithin}}{Return junctions whose
start coordinate matches, or is within, the
boundaries of the region requested.}

\item{\code{EndIsExactOrWithin}}{Return junctions whose
end coordinate matches, or is within, the
boudnaries of the region requested.}
}}

\examples{
query_jx(compilation = "gtex", regions = "CD99",
  coordinate_modifier = Coordinates$Exact)
}
\keyword{datasets}
