% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaptron.R
\name{exprs}
\alias{exprs}
\title{Utility function for formatting range and sample filters}
\usage{
exprs(..., frame = as.list(parent.frame()))
}
\arguments{
\item{...}{A list of expression to be converted to strings}

\item{frame}{An environment or list object}
}
\description{
This function makes it easier to compose sample and range filters for Snaptron queries.
Each expression is expected to be a boolean expression.
}
\examples{
a <- 10
exprs(sample_filters < a)
exprs(samples_count == a + 10)
}
