% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builder_wrappers.R
\name{from_url}
\alias{from_url}
\title{Constructs a \code{QueryBuilder} object from the given url}
\usage{
from_url(url)
}
\arguments{
\item{url}{a well-formed url preferably obtained from a call to the
\code{\link{uri_of_last_successful_request}} function}
}
\value{
Returns a \code{QueryBuilder} object with attributes set from the
parsed url.
}
\description{
Constructs a \code{QueryBuilder} object from the given url
}
\examples{
sb <- from_url("http://snaptron.cs.jhu.edu/gtex/snaptron?regions=CD99")
get_regions(sb)
get_compilation(sb)
}
