% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_level_functions.R
\name{junction_intersection}
\alias{junction_intersection}
\title{Get the intersection of junctions from 2 or more compilations
which are on the same reference}
\usage{
junction_intersection(...)
}
\arguments{
\item{...}{One or more SnaptronQueryBuilder objects}
}
\value{
A RangedSummarizedExperiment of junctions common across compilations
}
\description{
This function operates similar to the \code{junction_union()} function, i.e
it’s cross compilation and merging of the same junction from multiple
compilations will be handled exactly the same way. But instead
of every junction which appears in at least one compilation, only
the junctions which appear in \emph{every} compilation will be returned.
}
\examples{
sb1 <- SnaptronQueryBuilder$new()
sb1$from_url("http://snaptron.cs.jhu.edu/gtex/snaptron?regions=chr1:1879786-1879786&either=2&rfilter=strand:-")

sb2 <- SnaptronQueryBuilder$new()
sb2$from_url("http://snaptron.cs.jhu.edu/tcga/snaptron?regions=chr1:1879786-1879786&either=2&rfilter=strand:-")

junction_intersection(sb1, sb2)
}
\seealso{
\code{\link[=junction_union]{junction_union()}}
}
