% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_level_functions.R
\name{percent_spliced_in}
\alias{percent_spliced_in}
\title{Relative measure of splice variant usage, limited currently to
cassette exon splice variants}
\usage{
percent_spliced_in(inclusion_group1, inclusion_group2, exclusion_group,
  min_count = 20, group_names = NULL)
}
\arguments{
\item{inclusion_group1, inclusion_group2, exclusion_group}{A list of 1 or
more SnaptronQueryBuilder objects}

\item{min_count}{TODO: add description for min_count}

\item{group_names}{Optional vector of strings representing the group names}
}
\description{
Similar to the JIR, this calculates Percent Spliced In statistics
for the definition of 2 different groups: inclusion and exclusion.
Currently this function only supports the cassette exon use case.
}
\details{
Inclusion typically defines 2 basic queries, one for the junction
preceding the cassette exon, and the second for the junction following
the cassette exon.  The exclusion group contains one basic query
which defines the junction which skips the cassette exon.

The PSI itself is implemented as:\preformatted{mean(inclusion1, inclusion2) / (mean(inclusion1, inclusion2) + exclusion)```

where each term denotes the coverage of junctions that resulted
from the basic queries in that group in the current sample
}
}
\examples{
inclusion_group1 <- SnaptronQueryBuilder$new()
inclusion_group1 <- inclusion_group1$from_url("http://snaptron.cs.jhu.edu/srav2/snaptron?regions=chr1:94468008-94472172&exact=1&rfilter=strand:+")
inclusion_group2 <- SnaptronQueryBuilder$new()
inclusion_group2 <- inclusion_group2$from_url("http://snaptron.cs.jhu.edu/srav2/snaptron?regions=	&exact=1&rfilter=strand:+")
exclusion_group <- SnaptronQueryBuilder$new()
exclusion_group <- exclusion_group$from_url("http://snaptron.cs.jhu.edu/srav2/snaptron?regions=chr1:94468008-94475142&exact=1&rfilter=strand:+")
percent_spliced_in(list(inclusion_group1), list(inclusion_group2), list(exclusion_group))
}
