% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapcount.R
\docType{package}
\name{snapcount-package}
\alias{snapcount}
\alias{snapcount-package}
\title{snapcount: R/Bioconductor Package for interfacing with Snaptron for rapid quering of expression counts}
\description{
snapcount is a client interface to the Snaptron webservices
    which support querying by gene name or genomic region.
    Results include raw expression counts derived from alignment of RNA-seq samples
    and/or various summarized measures of expression across one or more regions/genes
    per-sample (e.g. percent spliced in).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/langmead-lab/snapr}
  \item Report bugs at \url{https://github.com/langmead-lab/snapr/issues}
}

}
