% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builder.R
\name{SnaptronQueryBuilder}
\alias{SnaptronQueryBuilder}
\title{A Reference Class for building Snaptron queries}
\description{
A Reference Class for building Snaptron queries

A Reference Class for building Snaptron queries
}
\section{Methods}{

\itemize{
\item{\code{compilation} Get or set query compilation. See
\code{\link{Compilation}} for details.}

\item{\code{regions} Get or set query region which can be either a chromosome
range or HUGO gene name.}

\item{\code{row_filters} Get or set query range filters. Range filters can be
set to a list or vector of either strings or R boolean expressions. If
filters are R boolean expressions then the right-hand-side of the R
expression should evaluate to either a string, float or whole number.
The expression will then be converted to a string before being sent out
to Snaptron.}

\item{\code{column_filters} Get or set query sample filters. When setting sample
filters the argument is expected to follow the sample rules as described
for \code{row_filters}.}

\item{\code{sids} Get or set query sample ids.}

\item{\code{coordinate_modifier} Get or set coordinate modifiers for query. See
\code{\link{Coordinates}} for details.}

\item{\code{query_jx} call \code{\link{query_jx}} function}

\item{\code{query_gene} call \code{\link{query_gene}} function}

\item{\code{query_exon} call \code{\link{query_exon}} function}

\item{\code{print} print query builder object}

\item{\code{from_url} use URL to instantiate SnaptronQueryBuilder object }
}
}

\examples{
sb <- SnaptronQueryBuilder$new()
sb$compilation("gtex")$regions("CD99")$query_jx()
sb$from_url("http://snaptron.cs.jhu.edu/gtex/snaptron?regions=chr1:1-100000")
sb$column_filters(SMTS == "Brain")

# The RHS of an R boolean expression can itself be an expression,
# but the expression has to evaluate to either a String, Float or Integer.
a <- 10
b <- 10
sb$row_filters(samples_count >= (a + b))
sb$query_jx(return_rse = FALSE)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{compilation}}{A single string containing the name of the Snaptron datasource}

\item{\code{regions}}{Either a list of 1 or more \code{HUGO} gene names \verb{(e.g. "BRCA1")} or a
GRanges-class object containing one or more genomic intervals \verb{(e.g. "chr1:1-1000")}.
Strand information is ignored.}

\item{\code{row_filters}}{A list of strings defining range-related contraints}

\item{\code{column_filters}}{A list of strings defining sample-related contraints}

\item{\code{sids}}{A list of rail_ids (integer sample IDs) to filter results on. Only
records which have been found in at least one of these samples will be returned.}

\item{\code{coordinate_modifier}}{Snaptron coordinate modifier enum. Variants include:

Coordinate$Exact - Contraints the results so that the start/end coordinates
match the start/end of the specifiied range.

Coordinate$Within - Contraints the results so that the coordinates are
within (inclusive) the specified range.

Coordinate$StartIsExactOrWithin - Constraints the results so that the start
coorindate matches or is within the boundaries of the specified range.

Coorindate$EndIsExactOrWithin - Contraints the results so that that the end.
coordinate matches or is within the boundaries of the specified range.

Coordinate$Overlaps - Contraints the results so that the coorindates overlaps the
specified range.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{compilation}}{A single string containing the name of the Snaptron datasource}

\item{\code{regions}}{Either a list of 1 or more \code{HUGO} gene names \verb{(e.g. "BRCA1")} or a
GRanges-class object containing one or more genomic intervals \verb{(e.g. "chr1:1-1000")}.
Strand information is ignored.}

\item{\code{row_filters}}{A list of strings defining range-related contraints}

\item{\code{column_filters}}{A list of strings defining sample-related contraints}

\item{\code{sids}}{A list of rail_ids (integer sample IDs) to filter results on. Only
records which have been found in at least one of these samples will be returned.}

\item{\code{coordinate_modifier}}{Snaptron coordinate modifier enum. Variants include:

Coordinate$Exact - Contraints the results so that the start/end coordinates
match the start/end of the specifiied range.

Coordinate$Within - Contraints the results so that the coordinates are
within (inclusive) the specified range.

Coordinate$StartIsExactOrWithin - Constraints the results so that the start
coorindate matches or is within the boundaries of the specified range.

Coorindate$EndIsExactOrWithin - Contraints the results so that that the end.
coordinate matches or is within the boundaries of the specified range.

Coordinate$Overlaps - Contraints the results so that the coorindates overlaps the
specified range.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SnaptronQueryBuilder$new()}}
\item \href{#method-compilation}{\code{SnaptronQueryBuilder$compilation()}}
\item \href{#method-regions}{\code{SnaptronQueryBuilder$regions()}}
\item \href{#method-row_filters}{\code{SnaptronQueryBuilder$row_filters()}}
\item \href{#method-column_filters}{\code{SnaptronQueryBuilder$column_filters()}}
\item \href{#method-sids}{\code{SnaptronQueryBuilder$sids()}}
\item \href{#method-coordinate_modifier}{\code{SnaptronQueryBuilder$coordinate_modifier()}}
\item \href{#method-query_jx}{\code{SnaptronQueryBuilder$query_jx()}}
\item \href{#method-query_exon}{\code{SnaptronQueryBuilder$query_exon()}}
\item \href{#method-query_gene}{\code{SnaptronQueryBuilder$query_gene()}}
\item \href{#method-from_url}{\code{SnaptronQueryBuilder$from_url()}}
\item \href{#method-print}{\code{SnaptronQueryBuilder$print()}}
\item \href{#method-clone}{\code{SnaptronQueryBuilder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnaptronQueryBuilder$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compilation"></a>}}
\subsection{Method \code{compilation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnaptronQueryBuilder$compilation(compilation = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-regions"></a>}}
\subsection{Method \code{regions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnaptronQueryBuilder$regions(regions = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-row_filters"></a>}}
\subsection{Method \code{row_filters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnaptronQueryBuilder$row_filters(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-column_filters"></a>}}
\subsection{Method \code{column_filters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnaptronQueryBuilder$column_filters(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sids"></a>}}
\subsection{Method \code{sids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnaptronQueryBuilder$sids(sids = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coordinate_modifier"></a>}}
\subsection{Method \code{coordinate_modifier()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnaptronQueryBuilder$coordinate_modifier(coordinate_modifier = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query_jx"></a>}}
\subsection{Method \code{query_jx()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnaptronQueryBuilder$query_jx(return_rse = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query_exon"></a>}}
\subsection{Method \code{query_exon()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnaptronQueryBuilder$query_exon(return_rse = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query_gene"></a>}}
\subsection{Method \code{query_gene()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnaptronQueryBuilder$query_gene(return_rse = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_url"></a>}}
\subsection{Method \code{from_url()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnaptronQueryBuilder$from_url(url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnaptronQueryBuilder$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SnaptronQueryBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
