% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaptron.R
\name{uri_of_last_successful_request}
\alias{uri_of_last_successful_request}
\title{Return the URI of the last successful request to Snaptron}
\usage{
uri_of_last_successful_request()
}
\value{
URI of last successful request to Snaptron or \code{NULL} if there have
not been any successful requests.
}
\description{
This function can be paired with the \code{from_url} method from the
SnaptronQueryBuilder class, allowing users to share sources of
data from Snaptron.
}
\examples{
query_jx(compilation = "gtex", genes_or_intervals = "CD99")
uri_of_last_successful_request()
}
