% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{Coordinates}
\alias{Coordinates}
\title{Snaptron Coordinate modifiers enum}
\format{An object of class \code{environment} of length 5.}
\usage{
Coordinates
}
\description{
Snaptron Coordinate modifiers enum
}
\keyword{datasets}
