% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_query_functions.R
\name{query_jx}
\alias{query_jx}
\alias{query_gene}
\alias{query_exon}
\title{Query Junctions/Genes/Exons}
\usage{
query_jx(compilation, genes_or_intervals, range_filters = NULL,
  sample_filters = NULL, sids = NULL, coordinate_modifier = NULL,
  return_rse = TRUE, split_by_region = FALSE)

query_gene(compilation, genes_or_intervals, range_filters = NULL,
  sample_filters = NULL, sids = NULL, coordinate_modifier = NULL,
  return_rse = TRUE, split_by_region = FALSE)

query_exon(compilation, genes_or_intervals, range_filters = NULL,
  sample_filters = NULL, sids = NULL, coordinate_modifier = NULL,
  return_rse = TRUE, split_by_region = FALSE)
}
\arguments{
\item{compilation}{A single string containing the name of the Snaptron datasource}

\item{genes_or_intervals}{Either a list of >=1 \code{HUGO} gene names \code{(e.g. "BRCA1")} or a
GRanges-class object containing one or more genomic intervals \code{(e.g. "chr1:1-1000")}.
Strand information is ignored.}

\item{range_filters}{A list of strings defining range-related contraints}

\item{sample_filters}{A list of strings defining sample-related contraints}

\item{sids}{A list of rail_ids (integer sample IDs) to filter results on. Only
records which have been found in at least one of these samples will be returned.}

\item{return_rse}{Should the query data be returned as a simple data frame or}

\item{either}{Contraints the results so that the start (either=1) or end
(either=2) coordinate matches or is within the boundaries of the specified range.}

\item{contains}{Constraints the results so that the coordinates within (inclusive)
the specified range.}

\item{exact}{Constraints the results so that the start/end coordinates
match the start/end of the specified range.}
}
\description{
Given one or more gene names or genomic range
intervals it will return a list of 0 or more genes, junctions, or exons
(depending on which query form is used) which overlap the ranges.
}
