% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaptron.R
\docType{data}
\name{SnaptronQueryBuilder}
\alias{SnaptronQueryBuilder}
\title{A Reference Class for building Snaptron queries}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SnaptronQueryBuilder
}
\description{
A Reference Class for building Snaptron queries
}
\examples{
sb <- SnaptronQueryBuilder$new()
sb$compilation("srav2")$genes("CD99")$query_jx()
}
\keyword{datasets}
