% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_query_functions.R
\docType{data}
\name{SnaptronQueryBuilder}
\alias{SnaptronQueryBuilder}
\title{A Reference Class for building Snaptron queries}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SnaptronQueryBuilder
}
\description{
A Reference Class for building Snaptron queries
}
\section{Methods}{

\code{compilation} Get or set snaptron compilation

\code{genes_or_intervals} Get or set either genes or intervals

\code{range_filters} Get or set range filters

\code{sample_filters} Get or set sample filters

\code{sids} Get or set Snaptron sample ids

\code{query_jx} call \code{\link{query_jx}} function

\code{query_gene} call \code{\link{query_gene}} function

\code{query_exon} call \code{\link{query_exon}} function

\code{query_coverage} call \code{\link{query_coverage}} function

\code{print} print query builder object

\code{from_url} use URL to instantiate SnaptronQueryBuilder object
}

\examples{
sb <- SnaptronQueryBuilder$new()
sb$compilation("gtex")$regions("CD99")$query_jx()
sb$from_url("http://snaptron.cs.jhu.edu/gtex/snaptron?regions=chr1:1-100000&rfilter=samples_count>:20&sfilter=SMTS:Brain")
}
\keyword{datasets}
