% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_builder.R
\docType{data}
\name{SnaptronQueryBuilder}
\alias{SnaptronQueryBuilder}
\title{A Reference Class for building Snaptron queries}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SnaptronQueryBuilder
}
\description{
A Reference Class for building Snaptron queries
}
\section{Fields}{

\describe{
\item{\code{compilation}}{A single string containing the name of the Snaptron datasource}

\item{\code{regions}}{Either a list of 1 or more \code{HUGO} gene names \code{(e.g. "BRCA1")} or a
GRanges-class object containing one or more genomic intervals \code{(e.g. "chr1:1-1000")}.
Strand information is ignored.}

\item{\code{row_filters}}{A list of strings defining range-related contraints}

\item{\code{column_filters}}{A list of strings defining sample-related contraints}

\item{\code{sids}}{A list of rail_ids (integer sample IDs) to filter results on. Only
records which have been found in at least one of these samples will be returned.}

\item{\code{coordinate_modifier}}{Snaptron coordinate modifier enum. Variants include:

Coordinate$Exact - Contraints the results so that the start/end coordinates
match the start/end of the specifiied range.

Coordinate$Within - Contraints the results so that the coordinates are
within (inclusive) the specified range.

Coordinate$StartIsExactOrWithin - Constraints the results so that the start
coorindate matches or is within the boundaries of the specified range.

Coorindate$EndIsExactOrWithin - Contraints the results so that that the end.
coordinate matches or is within the boundaries of the specified range.

Coordinate$Overlaps - Contraints the results so that the coorindates overlaps the
specified range.}
}}

\section{Methods}{

\describe{
\item{\code{compilation} Get or set query compilation. See
\code{\link{Compilation}} for details.}

\item{\code{regions} Get or set query region which can be either a chromosome
range or HUGO gene name.}

\item{\code{row_filters} Get or set query range filters. Range filters can be
set to a list or vector of either strings or R boolean expressions. If
filters are R boolean expressions then the right-hand-side of the R
expression should evaluate to either a string, float or whole number.
The expression will then be converted to a string before being sent out
to Snaptron.}

\item{\code{column_filters} Get or set query sample filters. When setting sample
filters the argument is expected to follow the sample rules as described
for \code{row_filters}.}

\item{\code{sids} Get or set query sample ids.}

\item{\code{coordinate_modifier} Get or set coordinate modifiers for query. See
\code{\link{Coordinates}} for details.}

\item{\code{query_jx} call \code{\link{query_jx}} function}

\item{\code{query_gene} call \code{\link{query_gene}} function}

\item{\code{query_exon} call \code{\link{query_exon}} function}

\item{\code{print} print query builder object}

\item{\code{from_url} use URL to instantiate SnaptronQueryBuilder object }}
}

\examples{
sb <- SnaptronQueryBuilder$new()
sb$compilation("gtex")$regions("CD99")$query_jx()
sb$from_url("http://snaptron.cs.jhu.edu/gtex/snaptron?regions=chr1:1-100000")
sb$sample_filters(SMTS == "Brain")

# The RHS of an R boolean expression can itself be an expression,
# but the expression has to evaluate to either a String, Float or Integer.
a <- 10
b <- 10
sb$range_filters(samples_count >= (a + b))
sb$query_jx(return_rse = FALSE)
}
\keyword{datasets}
